+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - add: eslint and change prettier config and start apply new coding style to root, test and public ts files
+- change : update_app.py to always use emojis
 
 ## 0.12.7 (2025-09-12)